# openapi.yaml
openapi: 3.1.0
info:
  title: StyleMatch API
  description: API for managing StyleMatch customers and their color profiles (color types, body shapes, and contrast levels).
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
  /ping:
    get:
      summary: Health check endpoint
      description: Returns "pong" to verify the application is running.
      responses:
        "200":
          description: Successful response
          content:
            text/plain:
              schema:
                type: string
                example: pong
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /customers:
    get:
      summary: Get all customers
      responses:
        "200":
          description: A list of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomerResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerRequest"
      responses:
        "201":
          description: Customer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /customers/{id}:
    get:
      summary: Get a customer by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Customer found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      summary: Update a customer
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerRequest"
      responses:
        "200":
          description: Customer updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      summary: Delete a customer
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Customer deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

components:
  responses:
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized access
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Access forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ColorType:
      type: string
      enum: [WINTER, SPRING, SUMMER, AUTUMN]
    BodyShape:
      type: string
      enum: [A, H, V, O, X]
    ContrastLevel:
      type: string
      enum: [LOW, MEDIUM, HIGH]
    Undertone:
      type: string
      enum: [WARM, COOL]

    CreateCustomerRequest:
      type: object
      required:
        - name
        - email
        - colorType
        - bodyShape
        - contrastLevel
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        colorType:
          $ref: "#/components/schemas/ColorType"
        bodyShape:
          $ref: "#/components/schemas/BodyShape"
        contrastLevel:
          $ref: "#/components/schemas/ContrastLevel"
        undertone:
          $ref: "#/components/schemas/Undertone"

    UpdateCustomerRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        colorType:
          $ref: "#/components/schemas/ColorType"
        bodyShape:
          $ref: "#/components/schemas/BodyShape"
        contrastLevel:
          $ref: "#/components/schemas/ContrastLevel"
        undertone:
          $ref: "#/components/schemas/Undertone"

    CustomerResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        colorType:
          $ref: "#/components/schemas/ColorType"
        bodyShape:
          $ref: "#/components/schemas/BodyShape"
        contrastLevel:
          $ref: "#/components/schemas/ContrastLevel"
        undertone:
          $ref: "#/components/schemas/Undertone"

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string